# cpu/sam0_common: SPI: don't perform DMA transfer for small buffers