--- a/cpu/samd21/periph/Kconfig.spi
+++ b/cpu/samd21/periph/Kconfig.spi
@@ -1,0 +1,1 @@
+source "$(RIOTCPU)/sam0_common/periph/Kconfig.spi"