--- a/cpu/saml1x/periph/Kconfig.spi
+++ b/cpu/saml1x/periph/Kconfig.spi
@@ -1,0 +1,1 @@
+source "$(RIOTCPU)/sam0_common/periph/Kconfig.spi"