--- a/drivers/periph_common/Kconfig.spi
+++ b/drivers/periph_common/Kconfig.spi
@@ -41,4 +41,7 @@ config MODULE_PERIPH_INIT_SPI_GPIO_MODE
     default y if MODULE_PERIPH_INIT
     depends on MODULE_PERIPH_SPI_GPIO_MODE
 
+# Include CPU specific configurations
+osource "$(RIOTCPU)/$(CPU)/periph/Kconfig.spi"
+
 endif # MODULE_PERIPH_SPI