<!--
The RIOT community cares a lot about code quality.
Therefore, before describing what your contribution is about, we would like
you to make sure that your modifications are compliant with the RIOT
coding conventions, see https://github.com/RIOT-OS/RIOT/wiki/Coding-conventions.
-->

### Contribution description

Setting up a DMA transfer can take longer than sending out a buffer byte by byte if the buffer is small.

DMA only shows advantages for large buffers, using it for every transfer will cause a net slowdown.

Since we did not come up with a good way to determine the treshold based on the SPI frequency, just use a fixed buffer for now so that DMA can be used without slowing things down overall.


### Testing procedure

The benchmark function of `tests/periph_spi_dma` make the effect apparent:

#### master

```
2021-10-07 13:36:58,396 # init 0 0 3
2021-10-07 13:36:58,406 # Trying to initialize SPI_DEV(0): mode: 0, clk: 3, cs_port: 0, cs_pin: 0
2021-10-07 13:36:58,408 # Note: Failed assertion (crash) means configuration not supported
2021-10-07 13:37:06,204 # bench
2021-10-07 13:37:06,208 # ### Running some benchmarks, all values in [us] ###
2021-10-07 13:37:06,211 # ### Test				Transfer time	user time
2021-10-07 13:37:06,212 # 
2021-10-07 13:37:06,226 #  1 - write 1000 times 1 byte:			11005	6011
2021-10-07 13:37:06,241 #  2 - write 1000 times 2 byte:			10829	5934
2021-10-07 13:37:06,418 #  3 - write 1000 times 100 byte:		173275	5918
2021-10-07 13:37:06,435 #  4 - write 1000 times 1 byte to register:	11713	6800
2021-10-07 13:37:06,452 #  5 - write 1000 times 2 byte to register:	12632	6666
2021-10-07 13:37:06,632 #  6 - write 1000 times 100 byte to register:	175692	6669
2021-10-07 13:37:06,647 #  7 - read 1000 times 2 byte:			10796	5800
2021-10-07 13:37:06,824 #  8 - read 1000 times 100 byte:		173243	5891
2021-10-07 13:37:06,841 #  9 - read 1000 times 2 byte from register:	12607	6632
2021-10-07 13:37:07,022 # 10 - read 1000 times 100 byte from register:	175668	6669
2021-10-07 13:37:07,037 # 11 - transfer 1000 times 2 byte:		11255	6258
2021-10-07 13:37:07,215 # 12 - transfer 1000 times 100 byte:		173258	5872
2021-10-07 13:37:07,232 # 13 - transfer 1000 times 2 byte to register:	12629	6633
2021-10-07 13:37:07,413 # 14 - transfer 1000 times 100 byte to register:175648	6641
2021-10-07 13:37:07,420 # 15 - acquire/release 1000 times:		3437	3442
2021-10-07 13:37:08,423 # -- - SUM:					1143687	91836
2021-10-07 13:37:08,423 # 
2021-10-07 13:37:08,425 # ### All runs complete ###
```

#### this patch

```
2021-10-07 13:38:34,643 # init 0 0 3
2021-10-07 13:38:34,649 # Trying to initialize SPI_DEV(0): mode: 0, clk: 3, cs_port: 0, cs_pin: 0
2021-10-07 13:38:34,655 # Note: Failed assertion (crash) means configuration not supported
2021-10-07 13:38:38,650 # bench
2021-10-07 13:38:38,654 # ### Running some benchmarks, all values in [us] ###
2021-10-07 13:38:38,658 # ### Test				Transfer time	user time
2021-10-07 13:38:38,658 # 
2021-10-07 13:38:38,665 #  1 - write 1000 times 1 byte:			3209	3213
2021-10-07 13:38:38,673 #  2 - write 1000 times 2 byte:			5012	5017
2021-10-07 13:38:38,851 #  3 - write 1000 times 100 byte:		173255	5886
2021-10-07 13:38:38,867 #  4 - write 1000 times 1 byte to register:	11712	6801
2021-10-07 13:38:38,884 #  5 - write 1000 times 2 byte to register:	12633	6675
2021-10-07 13:38:39,065 #  6 - write 1000 times 100 byte to register:	175704	6662
2021-10-07 13:38:39,073 #  7 - read 1000 times 2 byte:			5073	5079
2021-10-07 13:38:39,251 #  8 - read 1000 times 100 byte:		173248	5877
2021-10-07 13:38:39,268 #  9 - read 1000 times 2 byte from register:	12606	6640
2021-10-07 13:38:39,448 # 10 - read 1000 times 100 byte from register:	175625	6635
2021-10-07 13:38:39,457 # 11 - transfer 1000 times 2 byte:		5019	5024
2021-10-07 13:38:39,635 # 12 - transfer 1000 times 100 byte:		173225	5881
2021-10-07 13:38:39,652 # 13 - transfer 1000 times 2 byte to register:	12628	6633
2021-10-07 13:38:39,833 # 14 - transfer 1000 times 100 byte to register:175668	6640
2021-10-07 13:38:39,840 # 15 - acquire/release 1000 times:		3438	3442
2021-10-07 13:38:40,843 # -- - SUM:					1118055	86105
2021-10-07 13:38:40,843 # 
2021-10-07 13:38:40,845 # ### All runs complete ###

```

### Issues/PRs references

https://github.com/RIOT-OS/RIOT/pull/14261#issuecomment-644314287