@maribu 2021-10-01
lgtm. This could be exposed via KConfig, couldn't it?