@benpicco 2021-10-01
Yea although I think this is something the user shouldn't be getting too used to configure - it's after all the perfect example for a configuration option that only exists because the developer is lazy :wink: