@benpicco 2021-11-19
Ok! It still doesn't show up in `make BOARD=samr21-xpro menuconfig` in  `tests/periph_spi_dma` though