@leandrolanzieri 2021-11-19
> Ok! It still doesn't show up in `make BOARD=samr21-xpro menuconfig` in `tests/periph_spi_dma` though

As you are depending on `MODULE_` symbols, it will show up when `TEST_KCONFIG=1`.

There seems to be a conflict in `drivers/periph_common/Kconfig.spi`.